+++ static/app/components/metrics/equationSymbol.tsx
@@ -3,8 +3,7 @@ import {forwardRef} from 'react';
 import {hasMetricsNewInputs} from 'sentry/utils/metrics/features';
 import useOrganization from 'sentry/utils/useOrganization';
 
-import {QueryInputGroup} from './queryInputGroup';
-import {DeprecatedSymbol} from './styles';
+import {DeprecatedSymbol, Symbol} from './querySymbol';
 
 interface EquationSymbolProps extends React.ComponentProps<typeof DeprecatedSymbol> {
   equationId: number;
@@ -17,9 +16,7 @@ export function getEquationSymbol(equationId: number) {
 export const EquationSymbol = forwardRef<HTMLSpanElement, EquationSymbolProps>(
   function EquationSymbol({equationId, ...props}, ref) {
     const organization = useOrganization();
-    const Component = hasMetricsNewInputs(organization)
-      ? QueryInputGroup.Symbol
-      : DeprecatedSymbol;
+    const Component = hasMetricsNewInputs(organization) ? Symbol : DeprecatedSymbol;
     return (
       <Component ref={ref} {...props}>
         <span>
