+++ static/app/components/metrics/querySymbol.tsx
@@ -1,10 +1,10 @@
+import type {CSSProperties} from 'react';
 import {forwardRef} from 'react';
+import styled from '@emotion/styled';
 
+import {space} from 'sentry/styles/space';
 import {hasMetricsNewInputs} from 'sentry/utils/metrics/features';
 import useOrganization from 'sentry/utils/useOrganization';
-
-import {QueryInputGroup} from './queryInputGroup';
-import {DeprecatedSymbol} from './styles';
 
 const indexToChar = 'abcdefghijklmnopqrstuvwxyz';
 
@@ -18,7 +18,42 @@ export const getQuerySymbol = (index: number) => {
   return result;
 };
 
-interface QuerySymbolProps extends React.ComponentProps<typeof DeprecatedSymbol> {
+export const DeprecatedSymbol = styled('span')<{
+  cursor?: CSSProperties['cursor'];
+  isHidden?: boolean;
+}>`
+  display: flex;
+  width: 38px;
+  height: 38px;
+  line-height: 16px;
+  padding: ${space(0.5)};
+  justify-content: center;
+  align-items: center;
+  flex-shrink: 0;
+  border-radius: ${p => p.theme.borderRadius};
+  ${p => p.theme.fontWeightNormal};
+  color: ${p => p.theme.white};
+  font-size: 14px;
+  background: ${p => p.theme.purple300};
+  cursor: ${props => props.cursor ?? 'default'};
+  ${p =>
+    p.isHidden &&
+    `
+  background: ${p.theme.background};
+  color: ${p.theme.textColor};
+  border: 1px solid ${p.theme.border};
+  `}
+`;
+
+export const Symbol = styled(DeprecatedSymbol)`
+  color: ${p => p.theme.purple300};
+  border: 1px solid ${p => p.theme.purple200};
+  background: ${p => p.theme.purple100};
+  text-transform: uppercase;
+  font-weight: 500;
+`;
+
+interface QuerySymbolProps extends React.ComponentProps<typeof Symbol> {
   queryId: number;
 }
 
@@ -30,9 +65,7 @@ export const QuerySymbol = forwardRef<HTMLSpanElement, QuerySymbolProps>(
       return null;
     }
 
-    const Component = hasMetricsNewInputs(organization)
-      ? QueryInputGroup.Symbol
-      : DeprecatedSymbol;
+    const Component = hasMetricsNewInputs(organization) ? Symbol : DeprecatedSymbol;
 
     return (
       <Component ref={ref} {...props}>
